pyproject.toml
[project]
name = "mle-bench-agent"
version = "0.1.0"
description = "An MLE-Bench purple agent that solves Kaggle competitions using AIDE-style tree search"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk[http-server]>=0.3.20",
    "openai>=1.30.0",
    "uvicorn>=0.38.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.1",
]
